added second database

+++ app/Http/Controllers/Api/V1/UserController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers\Api\V1;
 
 use App\Models\User;
+use App\Models\PhoneNumber;
+use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\Api\V1\User\UserResource;
 use App\Http\Requests\Api\V1\User\RegisterRequest;
@@ -28,14 +30,27 @@ class UserController extends Controller
     {
         $validated = $request->validated();
         $validated['password'] = bcrypt($validated['password']);
-        $user = User::create($validated);
-        $user = $user->fresh();
-        $user->sendEmailVerificationNotification();
-        return response()->json(
-            [
-                'message' => "User registered successfully. Please verify your email",
-                'user' => new UserResource($user)
-            ], 201);
+        DB::connection('mysql')->beginTransaction();
+        DB::connection('safe_mysql')->beginTransaction();
+        try {
+            $user = User::create($validated);
+            $user = $user->fresh();
+            PhoneNumber::create(['user_id' => $user->id, 'number' => $validated['number']]);
+            $user->sendEmailVerificationNotification();
+            DB::connection('safe_mysql')->commit();
+            DB::connection('mysql')->commit();
+            return response()->json(
+                [
+                    'message' => "User registered successfully",
+                    'user' => new UserResource($user)
+                ],
+                201
+            );
+        } catch (\Exception $e) {
+            DB::connection('safe_mysql')->rollBack();
+            DB::connection('mysql')->rollBack();
+            throw $e;
+        }
     }
 
     /**
@@ -51,8 +66,19 @@ class UserController extends Controller
      */
     public function update(UpdateUserRequest $request, User $user)
     {
-        $user->update(['name' => $request->name]);
-        return response()->json(new UserResource($user));
+        DB::connection('mysql')->beginTransaction();
+        DB::connection('safe_mysql')->beginTransaction();
+        try {
+            $user->update(['name' => $request->name]);
+            PhoneNumber::updateOrCreate(['user_id' => $user->id], ['number' => $request->number]);
+            DB::connection('safe_mysql')->commit();
+            DB::connection('mysql')->commit();
+            return response()->json(new UserResource($user));
+        } catch (\Exception $e) {
+            DB::connection('safe_mysql')->rollBack();
+            DB::connection('mysql')->rollBack();
+            throw $e;
+        }
     }
 
     /**
@@ -60,22 +86,47 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
-        $user->delete();
-        return response()->json(null, 204);
+        DB::connection('mysql')->beginTransaction();
+        DB::connection('safe_mysql')->beginTransaction();
+        try {
+            PhoneNumber::where('user_id', $user->id)->delete();
+            $user->delete();
+            DB::connection('safe_mysql')->commit();
+            DB::connection('mysql')->commit();
+            return response()->json(null, 204);
+        } catch (\Exception $e) {
+            DB::connection('safe_mysql')->rollBack();
+            DB::connection('mysql')->rollBack();
+            throw $e;
+        }
     }
 
     public function register(RegisterRequest $request)
     {
         $validated = $request->validated();
         $validated['password'] = bcrypt($validated['password']);
-        $user = User::create($validated);
-        $user = $user->fresh();
-        $user->sendEmailVerificationNotification();
-        return response()->json(
-            [
-                'message' => "User registered successfully. please verify your email",
-                'user' => new UserResource($user),
-                'token' => $user->createToken('auth-token')->plainTextToken
-            ], 201);
+        DB::connection('mysql')->beginTransaction();
+        DB::connection('safe_mysql')->beginTransaction();
+        try {
+            $user = User::create($validated);
+            $user = $user->fresh();
+            PhoneNumber::create(['user_id' => $user->id, 'number' => $validated['number']]);
+            $user->sendEmailVerificationNotification();
+            DB::connection('safe_mysql')->commit();
+            DB::connection('mysql')->commit();
+            return response()->json(
+                [
+                    'message' => "User registered successfully. please verify your email",
+                    'user' => new UserResource($user),
+                    'token' => $user->createToken('auth-token')->plainTextToken
+                ],
+                201
+            );
+        } catch (\Exception $e) {
+            DB::connection('safe_mysql')->rollBack();
+            DB::connection('mysql')->rollBack();
+            throw $e;
+        }
     }
+
 }
